add posibillity to create city maps for whole region, scaling, different city frequency

+++ ProceduralTerrains/Assets/Scripts/TileProperties/TileProperties.cs
@@ -2,5 +2,5 @@ public abstract class TileProperties
 {
     public abstract void DrawGUI();
 
-    public abstract void CreateTile(MapGenerator mapGenerator, float xCoord, float yCoord);
+    public abstract void CreateTile(MapGenerator mapGenerator, int xCoord, int yCoord);
 }
